% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delta_dmca.R
\name{deltadmca.test}
\alias{deltadmca.test}
\title{Statistical test for Statistical test for DMCA cross-correlation coefficient.}
\usage{
deltadmca.test(x, y, k, rep)
}
\arguments{
\item{x}{A vector containing univariate time series.}

\item{y}{A vector containing univariate time series.}

\item{k}{An integer value indicating the boundary of the division \eqn{(N/k)}.
The smallest value of \eqn{k} is \eqn{4}.}

\item{rep}{An integer value indicating the number of repetitions.}
}
\value{
An list containing "timescale", "mean", "sd", "rho_before", "rho_affter", "deltarho", "CI_0.90", "CI_0.95", "CI_0.99".
}
\description{
This function performs the statistical test for Detrending moving-average cross-correlation coefficient from two univariate ARFIMA process.
}
\details{
This function include following measures: timescale, rho_before, rho_affter, deltarho
}
\examples{
x <- rnorm(1000)
y <-  rnorm(1000)
deltadmca.test(x,y,k=100,rep=10)

}
\references{
Guedes, et al. Statistical test for DCCA cross-correlation coefficient, Physica A, v.501, 134-140, 2018.

Guedes, et al. Statistical test for DCCA: Methods and data, Data in Brief, v. 18, 795-798, 2018.
}
